simplify pyenv setup (major speedup to bash init)
pyenv's install commands amount to adding an entry to the path, and
adding some functions for activate/deactivate. The whole point of pyenv
is to avoid needing to activate, so we don't care about the activate
aliases, and we already have an ensurepath helper defined; just use
that.

Also, take the existence of pyenv shims for python and pip as evidence
that pyenv is installed, to avoid a few command -v checks on the happy
path.